⚡ Bolt: Lazy load generation-specific save parsers (fixed)
Extracted detection logic to `src/engine/saveParser/utils/detection.ts` to prevent `INEFFECTIVE_DYNAMIC_IMPORT` warnings.
Refactored `parseSaveFile` in `src/engine/saveParser/index.ts` to use
dynamic imports for `parseGen1`, `parseGen2`, and `parseGen3`. Updated
`vite.config.ts` to output `saveParserGen1`, `saveParserGen2`, and
`saveParserGen3` as manual chunks, and adjusted `.bundlemonrc.json` limits.
