--- .jules/bolt/performance-lazy-load.md
@@ -1,1 +1,0 @@
-# Performance Optimization: Lazy Load Save Parsers\n\n- Used dynamic imports in `parseSaveFile` to lazily load `parseGen1`, `parseGen2`, and `parseGen3` depending on the detected save generation.\n- Configured Vite's `manualChunks` to break out `saveParserGen1`, `saveParserGen2`, and `saveParserGen3` into independent bundles.\n- Adjusted `.bundlemonrc.json` limits to accept the newly split chunk files.\n- Verified with `pnpm test:bundle` and E2E tests passing.\n- Added `// ⚡ Bolt:` inline comment documenting the optimization.\n\n
